Remove dynamic last data point
This change set removes the feature that updates the last data
point at each execution of RU. When merged, the last data point
will only be calculated after its full time range has elapsed.

As a side effect of this small change, there are other parts of
RU that are not needed any more, and are removed in this patch.
Those are:

1) Remove 'frequency' configuration option: its main functionality
was to define when the last data point should be updated. Not
needed any more.

2) As part of 1), remove last_exec_time global and corresponding
checks. This global was needed do check with the frequency, to
know when to execute updates to the last data point.

3) Likewise, there is no need of keeping a history file any more.
It was the file that stored the last_exec_time. So the history
file has also been removed.

4) Remove 'is_timeboxed=False' option: this is not strictly
needed, but this feature introduces more complexity in the code
and configuration and is not used today by any report in any
repository. This change makes all reports timeboxed.

5) Remove tests that were covering removed code. And fix tests
that would fail because of changes.

Bug: T131049
Change-Id: I855731f0e132bf5c537f84ee72d187c82841e9e3