Add 'execute' option to RU config
Sometimes a same query or script can be used to generate more than
one report. For instance, if we want to have the same report with
different granularities, or one with the whole history and another
with the data of the last 4 weeks.

The option execute=some_name allows to specify which query/script
should be executed instead of using the name of the report.

Bug: T131849
Change-Id: I58a9e9dd7c7353f7de1b386b2a372cfa692cab22