Add max_data_points option
This feature allows to specify a maximum size for the report using
the config option 'max_data_points=N', where N is an integer >= 1.
So, if N=5, the report will only hold the last 5 data points.
Note that this depends on the granularity, N=5 means last 5 weeks
for weekly granularity, and last 5 days for daily granularity, etc.

Bug: T131849
Change-Id: I99683d3ac3c54361d75d6fe4a1c72f3070312d30