Fix bug in get_date_threshold
After a report with max_data_points was trimmed, in the next
execution of reportupdater, the removed earlier dates would
have been recomputed (first error), and then not re-trimmed,
because the threshold was based on their date as opposed
of based on the latest date existing in the report. this
patch fixes both problems.

Bug: T131849
Change-Id: I33d488b0573c090443ab5046882ef0405ae0dc06