Add null values when query returns no results
Before, when the query or script returned no results (just header),
RU didn't update the report file for that time period, leaving a
blank in the report. This was on purpose and intended to give
another opportunity for the data to be backfilled when ready.

But in practice, when a query does not return any results at first,
it usually does not return any results in subsequent retrials.
So, the corresponding slot in the report file remains empty
forever, and the query is run every time RU is executed,
using unnecessary resources.

This change makes RU output a line with null values, when a query
or script returns no results. Thus, avoiding unnecessary
recomputations. If the data for a query is not ready at the time
RU triggers by default, the lag option should be used to delay
the execution by a given number of seconds.

Bug: T117537
Change-Id: Ic1790aed5df600759999b617d5900ad9a80085e0