Optionally record data to Graphite
Using the example config mentioned above, with the current query, the
actions and formats "None" are sent with "None" as the value.  I wasn't
sure if this was desired behavior or not, and if it makes sense to fix
it in the query or the record_to_graphite function.  Other than that,
ready to merge I think, I'll get mforns to review it.

Sample config can be found at: I7456a914f66ddb4b85b6af697a0b797a196e8763

Tip for testing: Run a static server on localhost:5000 and change the
graphite config to hit that to see what you'd be sending to graphite.

Bug: T150187
Change-Id: I820b088d2965579d32093dc36dccd0042f72ff77