reader.py: Get all report keys from defaults
Now all report keys can be specified in the "defaults" section, like the
"db" key. The value for a specific report will override the default
value. For many types of key, there is a global default value that is
used if the key isn't in the default either.

To avoid duplicating this logic, this patch adds a Reader.get_value
function.

Bug: T193171
Change-Id: I766612089884d5322c86209baa12c7ae2c2dccff