Use pid library to manage Pidfile
The pid library appears to be mature and should elegantly handle a
number of gnarly edge cases associated with pidfile management.

One thing to note is that pid is ASL-licensed. I can't figure out
how reportupdater is licensed so I'm not sure if that's okay. If ASL is
a problem, or if reviewers would prefer to avoid adding another
dependency, we could basically implement PidFile in reportupdater.

The pid library will try to put the pidfile in /var/run, or put it in a
temporary directory as a last resort. It's easy to change this.

Different from before, a job will now exit successfully if there's
already a job running with the same query folder.

Bug: T173604
Change-Id: I64b9152718effca852e74ec066bc330b7193b402